requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
faster-whisper==1.1.0
requests>=2.31
aiosqlite>=0.19.0
torch>=2.0
torchaudio>=2.0
python-pptx>=0.6.21
fpdf2>=2.7
